// ordergroove/types/paths/addresses.types.ts
// This code was automatically generated from an OpenAPI description.
// Do not edit this file. Edit the OpenAPI file instead.
// For more information, see https://github.com/pmcelhaney/counterfact/blob/main/docs/faq.md

import type { WideOperationArgument } from "../../counterfact-types/index.ts";
import type { OmitValueWhenNever } from "../../counterfact-types/index.ts";
import type { MaybePromise } from "../../counterfact-types/index.ts";
import type { COUNTERFACT_RESPONSE } from "../../counterfact-types/index.ts";
import type { Context } from "../../routes/_.context.ts";
import type { ResponseBuilderFactory } from "../../counterfact-types/index.ts";
import type { AddressPage } from "../components/schemas/AddressPage.js";
import type { AuthenticationFailed } from "../#/components/responses/AuthenticationFailed.js";

/**
 * Lists customer addresses. Application scope requires Bulk Operations
 * permission to return addresses for more than one customer. Source reviewed
 * 2026-08-02: https://developer.ordergroove.com/reference/addresses-list
 *
 */
export type addressesList = (
  $: OmitValueWhenNever<{
    query: addressesList_Query;
    querystring: never;
    path: never;
    headers: addressesList_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: AddressPage;
          };
        };
        examples: {};
      };
      403: AuthenticationFailed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type addressesList_Query = {
  updated_end?: string;
  /**
   * Customer ID; documented as application-scope only.
   */
  customer?: string;
  live?: boolean;
};

export type addressesList_Headers = {
  "x-api-key": string;
  Authorization: string;
};